requirements.txt
datasets
torch
more_itertools
tqdm
wandb
faiss-cpu
gensim
